+++ app/src/main/java/com/github/se/travelpouch/ui/dashboard/map/CameraUpdater.kt
@@ -1,0 +1,20 @@
+package com.github.se.travelpouch.ui.dashboard.map
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import com.github.se.travelpouch.model.activity.Activity
+import com.google.android.gms.maps.model.CameraPosition
+import com.google.android.gms.maps.model.LatLng
+import com.google.maps.android.compose.CameraPositionState
+
+@Composable
+fun CameraUpdater(listOfActivities: List<Activity>, cameraPositionState: CameraPositionState) {
+  LaunchedEffect(listOfActivities) {
+    if (listOfActivities.isNotEmpty()) {
+      val firstLocation = listOfActivities.first().location
+      cameraPositionState.position =
+          CameraPosition.fromLatLngZoom(
+              LatLng(firstLocation.latitude, firstLocation.longitude), 10f)
+    }
+  }
+}
